Fixed minor errors and added new command option

+++ resources/scripts/watchers/apexCodeWatcher.js
@@ -2,7 +2,7 @@ const vscode = require('vscode');
 const applicationContext = require('../core/applicationContext');
 const { ApexParser } = require('@aurahelper/languages').Apex;
 const Paths = require('../core/paths');
-const { FileWriter, PathUtils } = require('@aurahelper/core').FileSystem;
+const { FileWriter, PathUtils, FileChecker } = require('@aurahelper/core').FileSystem;
 
 class ApexCodeWatcher {
 
@@ -17,21 +17,38 @@ module.exports = ApexCodeWatcher;
 function registerApexClassesWatcher() {
     const classWatcher = vscode.workspace.createFileSystemWatcher("**/*.cls");
     classWatcher.onDidChange(async function (uri) {
-        ApexParser.saveClassData(uri.fsPath, Paths.getCompiledClassesFolder(), applicationContext.parserData).then(function (apexNode) {
-            applicationContext.parserData.userClassesData[apexNode.name.toLowerCase()] = apexNode;
-            analizeNodeErrors(apexNode);
-        });
+        if (FileChecker.isExists(uri.fsPath)) {
+            ApexParser.saveClassData(uri.fsPath, Paths.getCompiledClassesFolder(), applicationContext.parserData).then(function (apexNode) {
+                applicationContext.parserData.userClassesData[apexNode.name.toLowerCase()] = apexNode;
+                analizeNodeErrors(apexNode);
+            });
+        } else {
+            const fileName = PathUtils.getBasename(uri.fsPath);
+            const className = fileName.substring(0, fileName.indexOf('.'));
+            delete applicationContext.parserData.userClassesData[className.toLowerCase()];
+            if (FileChecker.isExists(Paths.getCompiledClassesFolder() + '/' + className + '.json'))
+                FileWriter.delete(Paths.getCompiledClassesFolder() + '/' + className + '.json');
+        }
     });
     classWatcher.onDidCreate(async function (uri) {
-        ApexParser.saveClassData(uri.fsPath, Paths.getCompiledClassesFolder(), applicationContext.parserData).then(function (apexNode) {
-            applicationContext.parserData.userClassesData[apexNode.name.toLowerCase()] = apexNode;
-            analizeNodeErrors(apexNode);
-        });
+        if (FileChecker.isExists(uri.fsPath)) {
+            ApexParser.saveClassData(uri.fsPath, Paths.getCompiledClassesFolder(), applicationContext.parserData).then(function (apexNode) {
+                applicationContext.parserData.userClassesData[apexNode.name.toLowerCase()] = apexNode;
+                analizeNodeErrors(apexNode);
+            });
+        } else {
+            const fileName = PathUtils.getBasename(uri.fsPath);
+            const className = fileName.substring(0, fileName.indexOf('.'));
+            delete applicationContext.parserData.userClassesData[className.toLowerCase()];
+            if (FileChecker.isExists(Paths.getCompiledClassesFolder() + '/' + className + '.json'))
+                FileWriter.delete(Paths.getCompiledClassesFolder() + '/' + className + '.json');
+        }
     });
     classWatcher.onDidDelete(async function (uri) {
         const fileName = PathUtils.getBasename(uri.fsPath);
         const className = fileName.substring(0, fileName.indexOf('.'));
-        FileWriter.delete(Paths.getCompiledClassesFolder() + '/' + className + '.json');
+        if (FileChecker.isExists(Paths.getCompiledClassesFolder() + '/' + className + '.json'))
+            FileWriter.delete(Paths.getCompiledClassesFolder() + '/' + className + '.json');
         delete applicationContext.parserData.userClassesData[className.toLowerCase()];
     });
 }
@@ -39,21 +56,38 @@ function registerApexClassesWatcher() {
 function registerApexTriggersWatcher() {
     const classWatcher = vscode.workspace.createFileSystemWatcher("**/*.trigger");
     classWatcher.onDidChange(async function (uri) {
-        ApexParser.saveClassData(uri.fsPath, Paths.getCompiledClassesFolder(), applicationContext.parserData).then(function (apexNode) {
-            applicationContext.parserData.userClassesData[apexNode.name.toLowerCase()] = apexNode;
-            analizeNodeErrors(apexNode);
-        });
+        if (FileChecker.isExists(uri.fsPath)) {
+            ApexParser.saveClassData(uri.fsPath, Paths.getCompiledClassesFolder(), applicationContext.parserData).then(function (apexNode) {
+                applicationContext.parserData.userClassesData[apexNode.name.toLowerCase()] = apexNode;
+                analizeNodeErrors(apexNode);
+            });
+        } else {
+            const fileName = PathUtils.getBasename(uri.fsPath);
+            const className = fileName.substring(0, fileName.indexOf('.'));
+            delete applicationContext.parserData.userClassesData[className.toLowerCase()];
+            if (FileChecker.isExists(Paths.getCompiledClassesFolder() + '/' + className + '.json'))
+                FileWriter.delete(Paths.getCompiledClassesFolder() + '/' + className + '.json');
+        }
     });
     classWatcher.onDidCreate(async function (uri) {
-        ApexParser.saveClassData(uri.fsPath, Paths.getCompiledClassesFolder(), applicationContext.parserData).then(function (apexNode) {
-            applicationContext.parserData.userClassesData[apexNode.name.toLowerCase()] = apexNode;
-            analizeNodeErrors(apexNode);
-        });
+        if (FileChecker.isExists(uri.fsPath)) {
+            ApexParser.saveClassData(uri.fsPath, Paths.getCompiledClassesFolder(), applicationContext.parserData).then(function (apexNode) {
+                applicationContext.parserData.userClassesData[apexNode.name.toLowerCase()] = apexNode;
+                analizeNodeErrors(apexNode);
+            });
+        } else {
+            const fileName = PathUtils.getBasename(uri.fsPath);
+            const className = fileName.substring(0, fileName.indexOf('.'));
+            delete applicationContext.parserData.userClassesData[className.toLowerCase()];
+            if (FileChecker.isExists(Paths.getCompiledClassesFolder() + '/' + className + '.json'))
+                FileWriter.delete(Paths.getCompiledClassesFolder() + '/' + className + '.json');
+        }
     });
     classWatcher.onDidDelete(async function (uri) {
         const fileName = PathUtils.getBasename(uri.fsPath);
         const className = fileName.substring(0, fileName.indexOf('.'));
-        FileWriter.delete(Paths.getCompiledClassesFolder() + '/' + className + '.json');
+        if (FileChecker.isExists(Paths.getCompiledClassesFolder() + '/' + className + '.json'))
+            FileWriter.delete(Paths.getCompiledClassesFolder() + '/' + className + '.json');
         delete applicationContext.parserData.userClassesData[className.toLowerCase()];
     });
 }
